remove downloadFile from bundle
License: MIT
